[ feat ] Q1 타입 에러 해결
