--- src/@components/emotion/Question2.jsx
@@ -1,134 +1,0 @@
-import { RefObject, useEffect, useRef } from 'react'
-import  styled  from 'styled-components'
-import { useRecoilState, useRecoilValue, useResetRecoilState, useSetRecoilState } from 'recoil';
-import { emotionData, stepData, userNameData } from '../../recoil/emotion';
-import { useNavigate } from 'react-router-dom';
-
-export default function Question2 () {
-  
-  const navigation = useNavigate()
-  
-  const [emotion, setEmotion] = useRecoilState(emotionData);
-  const userName = useRecoilValue(userNameData);
-  const setStep = useSetRecoilState(stepData);
-  
-  const QuestText = 
-  { 0: `님,
-오늘 하루는 어땠나요?
-한 줄로 이야기해 주세요.
-
-이게 사진의 이름이 될 거예요.` }
-  
-  
-  const moveToStep1 = () => {
-    setStep(1);
-  }
-
-    const moveToStep3 = () => {
-    setEmotion((prev) => ({ ...prev, title: todayRef.current.value }))
-    setStep(3);
-  }
-  
-  // const nameRef = useRef<HTMLInputElement>(null); TS
-  const todayRef = useRef(null); 
-
-
-  return (
-    <>
-        <St.AskBox>
-          <St.QuestionContainer>
-          {userName}{QuestText[0]}
-          </St.QuestionContainer>
-        </St.AskBox>
-        <St.AnswerName
-          placeholder='오늘 하루를 정리해주세요.'
-          type='text'
-          ref={todayRef} />
-        <St.ButtonContainer>
-          <St.PrevBtn onClick={moveToStep1}>이전</St.PrevBtn>
-          <St.NextBtn onClick={moveToStep3}>다음</St.NextBtn>
-      </St.ButtonContainer>
-      </>
-  )
-}
-
-const St = {
-
-  AskBox: styled.div`
-    background-color: rgba(256, 256, 256, 50%);
-    padding: 4rem 4rem; 
-    font-size: 1.6rem
-  `,
-
-  QuestionContainer: styled.article `
-    display: flex;
-    flex-direction: column;
-    text-align: center;
-    justify-content: center;
-    white-space: pre-line;
-  `,
-
-  AnswerName : styled.input`
-    width: 100%;
-    height: 50px;
-    margin: 55px 0px;
-
-    box-sizing: border-box;
-    padding: 10px;
-    
-    background: rgba(255, 255, 255, 0.5);
-
-    border-radius: 12px;
-    border: 1px solid #FFFFFF;
-    box-shadow: 0px 0px 4px 3px rgba(255, 255, 255, 0.2);
-  `,
-
-  ButtonContainer : styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 1.2rem;
-  `,
-
-  PrevBtn: styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    width: 10.4rem;
-    height: 6rem;
-
-    background: 
-      linear-gradient(88.06deg, rgba(255, 255, 255, 0.6) 25.59%, rgba(255, 255, 255, 0.2) 77.45%),
-      linear-gradient(180deg, rgba(255, 255, 255, 0.12) 0%, rgba(255, 255, 255, 0) 100%),
-      linear-gradient(272.47deg, rgba(255, 255, 255, 0.8) -4.42%, rgba(255, 255, 255, 0.2) 100.79%);
-    box-shadow: 0px 3px 10px 0px #00000026;
-      filter: drop-shadow(0px 3px 10px rgba(0, 0, 0, 0.15));
-
-    ${({ theme }) => theme.fonts.body1};
-    color: ${({ theme }) => theme.colors.gray3};
-
-    border-radius: 12px;
-    `,
-
-  NextBtn: styled.div `
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    width: 22rem;
-    height: 6rem;
-
-    background: 
-      linear-gradient(88.06deg, rgba(255, 255, 255, 0.6) 25.59%, rgba(255, 255, 255, 0.2) 77.45%),
-      linear-gradient(180deg, rgba(255, 255, 255, 0.12) 0%, rgba(255, 255, 255, 0) 100%),
-      linear-gradient(272.47deg, rgba(255, 255, 255, 0.8) -4.42%, rgba(255, 255, 255, 0.2) 100.79%);
-    box-shadow: 0px 3px 10px 0px #00000026;
-      filter: drop-shadow(0px 3px 10px rgba(0, 0, 0, 0.15));
-
-    ${({ theme }) => theme.fonts.body1};
-    color: ${({ theme }) => theme.colors.blue1};
-
-    border-radius: 12px;
-
-  `
-}
